Fix nav-stack reset + voice-loop bugs found in E2E investigation
- JetBottomBar: tab taps land on the tab root (no saveState/restoreState),
  stale feature screens can no longer resurrect on tab switch
- MainActivity: latch the startup branch after first Loaded(true) so
  transient pref re-emissions can't recreate JetSetterApp/NavController
  (root cause of the silent reset-to-Home during hands-free voice)
- JetSetterApp: debug-only JetNav destination logging for future forensics
- VoiceInput: silence-length extras cast to Int (recognizer threw
  ClassCastException internally and ignored the 1.2s EOU tuning)
- VoiceLoopController: generation guard prevents stale onDone/retry from
  opening a second recognizer session during teardown

Also includes docs/PLAY_RELEASE_HANDOFF.md — release-state handoff doc
authored by the parallel Play-setup session, found in the working tree.

236 unit tests green; emulator re-verify: 3/3 hands-free cycles with no
nav reset, tab-root behavior confirmed, zero recognizer exceptions.

+++ app/src/main/java/com/jetsetter/pro/core/voice/VoiceInput.kt
@@ -93,13 +93,15 @@ class VoiceInput(private val context: Context) {
             )
             putExtra(RecognizerIntent.EXTRA_PARTIAL_RESULTS, onPartial != null)
             if (preferOffline) putExtra(RecognizerIntent.EXTRA_PREFER_OFFLINE, true)
+            // The recognizer service reads these extras as Int — a Long value makes it throw
+            // ClassCastException internally and fall back to 0, so convert at the boundary.
             putExtra(
                 RecognizerIntent.EXTRA_SPEECH_INPUT_COMPLETE_SILENCE_LENGTH_MILLIS,
-                completeSilenceMillis,
+                completeSilenceMillis.toInt(),
             )
             putExtra(
                 RecognizerIntent.EXTRA_SPEECH_INPUT_POSSIBLY_COMPLETE_SILENCE_LENGTH_MILLIS,
-                completeSilenceMillis,
+                completeSilenceMillis.toInt(),
             )
         }
